@@ -84,29 +95,24 @@
     "copy-dir": "^1.2.0",
     "cross-env": "^7.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-playwright": "^0.12.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.2",
+    "eslint-plugin-playwright": "^0.16.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "globals": "^13.20.0",
     "husky": "^8.0.3",
     "jest": "^27.4.7",
     "lint-staged": "^13.2.2",
     "live-server": "^1.2.1",
     "mkdirp": "^3.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.8",
     "rimraf": "^3.0.0",
-    "rollup": "^1.23.1",
-    "rollup-plugin-async": "^1.2.0",
-    "rollup-plugin-buble": "^0.19.8",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-uglify": "^6.0.4",
+    "rollup": "^3.25.3",
     "serve-handler": "^6.1.2",
-    "stylus": "^0.59.0",
+    "stylus": "^0.60.0",
     "vue2": "npm:vue@^2.6.12",
     "vue3": "npm:vue@^3.0.0",
     "xhr-mock": "^2.5.1"
