@@ -3,11 +3,10 @@ import { Render } from './render/index.js';
 import { Fetch } from './fetch/index.js';
 import { Events } from './event/index.js';
 import { VirtualRoutes } from './virtual-routes/index.js';
-import initGlobalAPI from './global-api.js';
 
 import config from './config.js';
-import { isFn } from './util/core';
-import { Lifecycle } from './init/lifecycle';
+import { isFn } from './util/core.js';
+import { Lifecycle } from './init/lifecycle.js';
 
 /** @typedef {new (...args: any[]) => any} Constructor */
 
@@ -16,10 +15,10 @@ export class Docsify extends Fetch(
   // eslint-disable-next-line new-cap
   Events(Render(VirtualRoutes(Router(Lifecycle(Object)))))
 ) {
+  config = config(this);
+
   constructor() {
     super();
-
-    this.config = config(this);
 
     this.initLifecycle(); // Init hooks
     this.initPlugin(); // Install plugins
@@ -32,7 +31,7 @@ export class Docsify extends Fetch(
   }
 
   initPlugin() {
-    [].concat(this.config.plugins).forEach(fn => {
+    this.config.plugins.forEach(fn => {
       try {
         isFn(fn) && fn(this._lifecycle, this);
       } catch (err) {
@@ -46,8 +45,3 @@ export class Docsify extends Fetch(
     });
   }
 }
-
-/**
- * Global API
- */
-initGlobalAPI();
