@@ -1,8 +1,17 @@
-import { documentReady } from './util/dom';
-import { Docsify } from './Docsify';
+import { documentReady } from './util/dom.js';
+import { Docsify } from './Docsify.js';
+import initGlobalAPI from './global-api.js';
+
+// TODO This global API and auto-running Docsify will be deprecated, and removed
+// in a major release. Instead we'll tell users to use `new Docsify()` to create
+// and manage their instance(s).
+
+/**
+ * Global API
+ */
+initGlobalAPI();
 
 /**
  * Run Docsify
  */
-// eslint-disable-next-line no-unused-vars
-documentReady(_ => new Docsify());
+documentReady(() => new Docsify());
