@@ -2,8 +2,8 @@ export const taskListItemCompiler = ({ renderer }) =>
   (renderer.listitem = text => {
     const isTaskItem = /^(<input.*type="checkbox"[^>]*>)/.test(text);
     const html = isTaskItem
-      ? `<li class="task-list-item"><label>${text}</label></li>`
-      : `<li>${text}</li>`;
+      ? /* html */ `<li class="task-list-item"><label>${text}</label></li>`
+      : /* html */ `<li>${text}</li>`;
 
     return html;
   });
