@@ -18,7 +18,7 @@ export function genTree(toc, maxLevel) {
     }
 
     if (last[len]) {
-      last[len].children = (last[len].children || []).concat(headline);
+      last[len].children = [...(last[len].children || []), headline];
     } else {
       headlines.push(headline);
     }
