@@ -1,243 +1,14 @@
 /* eslint-disable no-unused-vars */
 import tinydate from 'tinydate';
-import * as dom from '../util/dom';
-import cssVars from '../util/polyfill/css-vars';
-import { getAndActive, sticky } from '../event/sidebar';
-import { getPath, isAbsolutePath } from '../router/util';
-import { isMobile, inBrowser } from '../util/env';
-import { isPrimitive, merge } from '../util/core';
-import { scrollActiveSidebar } from '../event/scroll';
-import { Compiler } from './compiler';
-import * as tpl from './tpl';
-import { prerenderEmbed } from './embed';
+import * as dom from '../util/dom.js';
+import { getPath, isAbsolutePath } from '../router/util.js';
+import { isMobile, inBrowser } from '../util/env.js';
+import { isPrimitive } from '../util/core.js';
+import { Compiler } from './compiler.js';
+import * as tpl from './tpl.js';
+import { prerenderEmbed } from './embed.js';
 
-let vueGlobalData;
-
-function executeScript() {
-  const script = dom
-    .findAll('.markdown-section>script')
-    .filter(s => !/template/.test(s.type))[0];
-  if (!script) {
-    return false;
-  }
-
-  const code = script.innerText.trim();
-  if (!code) {
-    return false;
-  }
-
-  new Function(code)();
-}
-
-function formatUpdated(html, updated, fn) {
-  updated =
-    typeof fn === 'function'
-      ? fn(updated)
-      : typeof fn === 'string'
-      ? tinydate(fn)(new Date(updated))
-      : updated;
-
-  return html.replace(/{docsify-updated}/g, updated);
-}
-
-function renderMain(html) {
-  const docsifyConfig = this.config;
-  const markdownElm = dom.find('.markdown-section');
-  const vueVersion =
-    'Vue' in window &&
-    window.Vue.version &&
-    Number(window.Vue.version.charAt(0));
-
-  const isMountedVue = elm => {
-    const isVue2 = Boolean(elm.__vue__ && elm.__vue__._isVue);
-    const isVue3 = Boolean(elm._vnode && elm._vnode.__v_skip);
-
-    return isVue2 || isVue3;
-  };
-
-  if (!html) {
-    html = '<h1>404 - Not found</h1>';
-  }
-
-  if ('Vue' in window) {
-    const mountedElms = dom
-      .findAll('.markdown-section > *')
-      .filter(elm => isMountedVue(elm));
-
-    // Destroy/unmount existing Vue instances
-    for (const mountedElm of mountedElms) {
-      if (vueVersion === 2) {
-        mountedElm.__vue__.$destroy();
-      } else if (vueVersion === 3) {
-        mountedElm.__vue_app__.unmount();
-      }
-    }
-  }
-
-  this._renderTo(markdownElm, html);
-
-  // Render sidebar with the TOC
-  !docsifyConfig.loadSidebar && this._renderSidebar();
-
-  // Execute markdown <script>
-  if (
-    docsifyConfig.executeScript ||
-    ('Vue' in window && docsifyConfig.executeScript !== false)
-  ) {
-    executeScript();
-  }
-
-  // Handle Vue content not mounted by markdown <script>
-  if ('Vue' in window) {
-    const vueMountData = [];
-    const vueComponentNames = Object.keys(docsifyConfig.vueComponents || {});
-
-    // Register global vueComponents
-    if (vueVersion === 2 && vueComponentNames.length) {
-      vueComponentNames.forEach(name => {
-        const isNotRegistered = !window.Vue.options.components[name];
-
-        if (isNotRegistered) {
-          window.Vue.component(name, docsifyConfig.vueComponents[name]);
-        }
-      });
-    }
-
-    // Store global data() return value as shared data object
-    if (
-      !vueGlobalData &&
-      docsifyConfig.vueGlobalOptions &&
-      typeof docsifyConfig.vueGlobalOptions.data === 'function'
-    ) {
-      vueGlobalData = docsifyConfig.vueGlobalOptions.data();
-    }
-
-    // vueMounts
-    vueMountData.push(
-      ...Object.keys(docsifyConfig.vueMounts || {})
-        .map(cssSelector => [
-          dom.find(markdownElm, cssSelector),
-          docsifyConfig.vueMounts[cssSelector],
-        ])
-        .filter(([elm, vueConfig]) => elm)
-    );
-
-    // Template syntax, vueComponents, vueGlobalOptions
-    if (docsifyConfig.vueGlobalOptions || vueComponentNames.length) {
-      const reHasBraces = /{{2}[^{}]*}{2}/;
-      // Matches Vue full and shorthand syntax as attributes in HTML tags.
-      //
-      // Full syntax examples:
-      // v-foo, v-foo[bar], v-foo-bar, v-foo:bar-baz.prop
-      //
-      // Shorthand syntax examples:
-      // @foo, @foo.bar, @foo.bar.baz, @[foo], :foo, :[foo]
-      //
-      // Markup examples:
-      // <div v-html>{{ html }}</div>
-      // <div v-text="msg"></div>
-      // <div v-bind:text-content.prop="text">
-      // <button v-on:click="doThis"></button>
-      // <button v-on:click.once="doThis"></button>
-      // <button v-on:[event]="doThis"></button>
-      // <button @click.stop.prevent="doThis">
-      // <a :href="url">
-      // <a :[key]="url">
-      const reHasDirective = /<[^>/]+\s([@:]|v-)[\w-:.[\]]+[=>\s]/;
-
-      vueMountData.push(
-        ...dom
-          .findAll('.markdown-section > *')
-          // Remove duplicates
-          .filter(elm => !vueMountData.some(([e, c]) => e === elm))
-          // Detect Vue content
-          .filter(elm => {
-            const isVueMount =
-              // is a component
-              elm.tagName.toLowerCase() in
-                (docsifyConfig.vueComponents || {}) ||
-              // has a component(s)
-              elm.querySelector(vueComponentNames.join(',') || null) ||
-              // has curly braces
-              reHasBraces.test(elm.outerHTML) ||
-              // has content directive
-              reHasDirective.test(elm.outerHTML);
-
-            return isVueMount;
-          })
-          .map(elm => {
-            // Clone global configuration
-            const vueConfig = merge({}, docsifyConfig.vueGlobalOptions || {});
-
-            // Replace vueGlobalOptions data() return value with shared data object.
-            // This provides a global store for all Vue instances that receive
-            // vueGlobalOptions as their configuration.
-            if (vueGlobalData) {
-              vueConfig.data = function () {
-                return vueGlobalData;
-              };
-            }
-
-            return [elm, vueConfig];
-          })
-      );
-    }
-
-    // Mount
-    for (const [mountElm, vueConfig] of vueMountData) {
-      const isVueAttr = 'data-isvue';
-      const isSkipElm =
-        // Is an invalid tag
-        mountElm.matches('pre, script') ||
-        // Is a mounted instance
-        isMountedVue(mountElm) ||
-        // Has mounted instance(s)
-        mountElm.querySelector(`[${isVueAttr}]`);
-
-      if (!isSkipElm) {
-        mountElm.setAttribute(isVueAttr, '');
-
-        if (vueVersion === 2) {
-          vueConfig.el = undefined;
-          new window.Vue(vueConfig).$mount(mountElm);
-        } else if (vueVersion === 3) {
-          const app = window.Vue.createApp(vueConfig);
-
-          // Register global vueComponents
-          vueComponentNames.forEach(name => {
-            const config = docsifyConfig.vueComponents[name];
-
-            app.component(name, config);
-          });
-
-          app.mount(mountElm);
-        }
-      }
-    }
-  }
-}
-
-function renderNameLink(vm) {
-  const el = dom.getNode('.app-name-link');
-  const nameLink = vm.config.nameLink;
-  const path = vm.route.path;
-
-  if (!el) {
-    return;
-  }
-
-  if (isPrimitive(vm.config.nameLink)) {
-    el.setAttribute('href', nameLink);
-  } else if (typeof nameLink === 'object') {
-    const match = Object.keys(nameLink).filter(
-      key => path.indexOf(key) > -1
-    )[0];
-
-    el.setAttribute('href', nameLink[match]);
-  }
-}
-
-/** @typedef {import('../Docsify').Constructor} Constructor */
+/** @typedef {import('../Docsify.js').Constructor} Constructor */
 
 /**
  * @template {!Constructor} T
@@ -245,6 +16,232 @@ function renderNameLink(vm) {
  */
 export function Render(Base) {
   return class Render extends Base {
+    #vueGlobalData;
+
+    #executeScript() {
+      const script = dom
+        .findAll('.markdown-section>script')
+        .filter(s => !/template/.test(s.type))[0];
+      if (!script) {
+        return false;
+      }
+
+      const code = script.innerText.trim();
+      if (!code) {
+        return false;
+      }
+
+      new Function(code)();
+    }
+
+    #formatUpdated(html, updated, fn) {
+      updated =
+        typeof fn === 'function'
+          ? fn(updated)
+          : typeof fn === 'string'
+          ? tinydate(fn)(new Date(updated))
+          : updated;
+
+      return html.replace(/{docsify-updated}/g, updated);
+    }
+
+    #renderMain(html) {
+      const docsifyConfig = this.config;
+      const markdownElm = dom.find('.markdown-section');
+      const vueVersion =
+        'Vue' in window &&
+        window.Vue.version &&
+        Number(window.Vue.version.charAt(0));
+
+      const isMountedVue = elm => {
+        const isVue2 = Boolean(elm.__vue__ && elm.__vue__._isVue);
+        const isVue3 = Boolean(elm._vnode && elm._vnode.__v_skip);
+
+        return isVue2 || isVue3;
+      };
+
+      if (!html) {
+        html = /* html */ `<h1>404 - Not found</h1>`;
+      }
+
+      if ('Vue' in window) {
+        const mountedElms = dom
+          .findAll('.markdown-section > *')
+          .filter(elm => isMountedVue(elm));
+
+        // Destroy/unmount existing Vue instances
+        for (const mountedElm of mountedElms) {
+          if (vueVersion === 2) {
+            mountedElm.__vue__.$destroy();
+          } else if (vueVersion === 3) {
+            mountedElm.__vue_app__.unmount();
+          }
+        }
+      }
+
+      this._renderTo(markdownElm, html);
+
+      // Render sidebar with the TOC
+      !docsifyConfig.loadSidebar && this._renderSidebar();
+
+      // Execute markdown <script>
+      if (
+        docsifyConfig.executeScript ||
+        ('Vue' in window && docsifyConfig.executeScript !== false)
+      ) {
+        this.#executeScript();
+      }
+
+      // Handle Vue content not mounted by markdown <script>
+      if ('Vue' in window) {
+        const vueMountData = [];
+        const vueComponentNames = Object.keys(
+          docsifyConfig.vueComponents || {}
+        );
+
+        // Register global vueComponents
+        if (vueVersion === 2 && vueComponentNames.length) {
+          vueComponentNames.forEach(name => {
+            const isNotRegistered = !window.Vue.options.components[name];
+
+            if (isNotRegistered) {
+              window.Vue.component(name, docsifyConfig.vueComponents[name]);
+            }
+          });
+        }
+
+        // Store global data() return value as shared data object
+        if (
+          !this.#vueGlobalData &&
+          docsifyConfig.vueGlobalOptions &&
+          typeof docsifyConfig.vueGlobalOptions.data === 'function'
+        ) {
+          this.#vueGlobalData = docsifyConfig.vueGlobalOptions.data();
+        }
+
+        // vueMounts
+        vueMountData.push(
+          ...Object.keys(docsifyConfig.vueMounts || {})
+            .map(cssSelector => [
+              dom.find(markdownElm, cssSelector),
+              docsifyConfig.vueMounts[cssSelector],
+            ])
+            .filter(([elm, vueConfig]) => elm)
+        );
+
+        // Template syntax, vueComponents, vueGlobalOptions
+        if (docsifyConfig.vueGlobalOptions || vueComponentNames.length) {
+          const reHasBraces = /{{2}[^{}]*}{2}/;
+          // Matches Vue full and shorthand syntax as attributes in HTML tags.
+          //
+          // Full syntax examples:
+          // v-foo, v-foo[bar], v-foo-bar, v-foo:bar-baz.prop
+          //
+          // Shorthand syntax examples:
+          // @foo, @foo.bar, @foo.bar.baz, @[foo], :foo, :[foo]
+          //
+          // Markup examples:
+          // <div v-html>{{ html }}</div>
+          // <div v-text="msg"></div>
+          // <div v-bind:text-content.prop="text">
+          // <button v-on:click="doThis"></button>
+          // <button v-on:click.once="doThis"></button>
+          // <button v-on:[event]="doThis"></button>
+          // <button @click.stop.prevent="doThis">
+          // <a :href="url">
+          // <a :[key]="url">
+          const reHasDirective = /<[^>/]+\s([@:]|v-)[\w-:.[\]]+[=>\s]/;
+
+          vueMountData.push(
+            ...dom
+              .findAll('.markdown-section > *')
+              // Remove duplicates
+              .filter(elm => !vueMountData.some(([e, c]) => e === elm))
+              // Detect Vue content
+              .filter(elm => {
+                const isVueMount =
+                  // is a component
+                  elm.tagName.toLowerCase() in
+                    (docsifyConfig.vueComponents || {}) ||
+                  // has a component(s)
+                  elm.querySelector(vueComponentNames.join(',') || null) ||
+                  // has curly braces
+                  reHasBraces.test(elm.outerHTML) ||
+                  // has content directive
+                  reHasDirective.test(elm.outerHTML);
+
+                return isVueMount;
+              })
+              .map(elm => {
+                // Clone global configuration
+                const vueConfig = {
+                  ...docsifyConfig.vueGlobalOptions,
+                };
+                // Replace vueGlobalOptions data() return value with shared data object.
+                // This provides a global store for all Vue instances that receive
+                // vueGlobalOptions as their configuration.
+                if (this.#vueGlobalData) {
+                  vueConfig.data = () => this.#vueGlobalData;
+                }
+
+                return [elm, vueConfig];
+              })
+          );
+        }
+
+        // Mount
+        for (const [mountElm, vueConfig] of vueMountData) {
+          const isVueAttr = 'data-isvue';
+          const isSkipElm =
+            // Is an invalid tag
+            mountElm.matches('pre, script') ||
+            // Is a mounted instance
+            isMountedVue(mountElm) ||
+            // Has mounted instance(s)
+            mountElm.querySelector(`[${isVueAttr}]`);
+
+          if (!isSkipElm) {
+            mountElm.setAttribute(isVueAttr, '');
+
+            if (vueVersion === 2) {
+              vueConfig.el = undefined;
+              new window.Vue(vueConfig).$mount(mountElm);
+            } else if (vueVersion === 3) {
+              const app = window.Vue.createApp(vueConfig);
+
+              // Register global vueComponents
+              vueComponentNames.forEach(name => {
+                const config = docsifyConfig.vueComponents[name];
+
+                app.component(name, config);
+              });
+
+              app.mount(mountElm);
+            }
+          }
+        }
+      }
+    }
+
+    #renderNameLink(vm) {
+      const el = dom.getNode('.app-name-link');
+      const nameLink = vm.config.nameLink;
+      const path = vm.route.path;
+
+      if (!el) {
+        return;
+      }
+
+      if (isPrimitive(vm.config.nameLink)) {
+        el.setAttribute('href', nameLink);
+      } else if (typeof nameLink === 'object') {
+        const match = Object.keys(nameLink).filter(
+          key => path.indexOf(key) > -1
+        )[0];
+
+        el.setAttribute('href', nameLink[match]);
+      }
+    }
     _renderTo(el, content, replace) {
       const node = dom.getNode(el);
       if (node) {
@@ -271,7 +268,12 @@ export function Render(Base) {
       }
 
       this._renderTo('.sidebar-nav', this.compiler.sidebar(text, maxLevel));
-      const activeEl = getAndActive(this.router, '.sidebar-nav', true, true);
+      const activeEl = this.__getAndActive(
+        this.router,
+        '.sidebar-nav',
+        true,
+        true
+      );
       if (loadSidebar && activeEl) {
         activeEl.parentNode.innerHTML +=
           this.compiler.subSidebar(subMaxLevel) || '';
@@ -287,7 +289,7 @@ export function Render(Base) {
     _bindEventOnRendered(activeEl) {
       const { autoHeader } = this.config;
 
-      scrollActiveSidebar(this.router);
+      this.__scrollActiveSidebar(this.router);
 
       if (autoHeader && activeEl) {
         const main = dom.getNode('#main');
@@ -303,20 +305,20 @@ export function Render(Base) {
     _renderNav(text) {
       text && this._renderTo('nav', this.compiler.compile(text));
       if (this.config.loadNavbar) {
-        getAndActive(this.router, 'nav');
+        this.__getAndActive(this.router, 'nav');
       }
     }
 
     _renderMain(text, opt = {}, next) {
       if (!text) {
-        return renderMain.call(this, text);
+        return this.#renderMain(text);
       }
 
       this.callHook('beforeEach', text, result => {
         let html;
         const callback = () => {
           if (opt.updatedAt) {
-            html = formatUpdated(
+            html = this.#formatUpdated(
               html,
               opt.updatedAt,
               this.config.formatUpdated
@@ -324,7 +326,7 @@ export function Render(Base) {
           }
 
           this.callHook('afterEach', html, hookData => {
-            renderMain.call(this, hookData);
+            this.#renderMain(hookData);
             next();
           });
         };
@@ -388,12 +390,12 @@ export function Render(Base) {
       }
 
       this._renderTo('.cover-main', html);
-      sticky();
+      this.__sticky();
     }
 
     _updateRender() {
       // Render name link
-      renderNameLink(this);
+      this.#renderNameLink(this);
     }
 
     initRender() {
@@ -458,8 +460,6 @@ export function Render(Base) {
         dom.$.head.appendChild(
           dom.create('div', tpl.theme(config.themeColor)).firstElementChild
         );
-        // Polyfll
-        cssVars(config.themeColor);
       }
 
       this._updateRender();
