@@ -1,18 +1,15 @@
-import { isExternal, noop } from '../../util/core';
-import { on } from '../../util/dom';
-import { endsWith } from '../../util/str';
-import { parseQuery, cleanPath, replaceSlug } from '../util';
-import { History } from './base';
+import { isExternal, noop } from '../../util/core.js';
+import { on } from '../../util/dom.js';
+import { parseQuery, cleanPath, replaceSlug } from '../util.js';
+import { History } from './base.js';
 
 function replaceHash(path) {
   const i = location.href.indexOf('#');
   location.replace(location.href.slice(0, i >= 0 ? i : 0) + '#' + path);
 }
+
 export class HashHistory extends History {
-  constructor(config) {
-    super(config);
-    this.mode = 'hash';
-  }
+  mode = 'hash';
 
   getBasePath() {
     const path = window.location.pathname || '';
@@ -23,7 +20,7 @@ export class HashHistory extends History {
     // prevents the `/index.html` part of the URI from being
     // remove during routing.
     // See here: https://github.com/docsifyjs/docsify/pull/1372
-    const basePath = endsWith(path, '.html')
+    const basePath = path.endsWith('.html')
       ? path + '#/' + base
       : path + '/' + base;
     return /^(\/|https?:)/g.test(base) ? base : cleanPath(basePath);
