@@ -1,8 +1,7 @@
-import { supportsPushState } from '../util/env';
-import * as dom from '../util/dom';
-import { noop } from '../util/core';
-import { HashHistory } from './history/hash';
-import { HTML5History } from './history/html5';
+import * as dom from '../util/dom.js';
+import { noop } from '../util/core.js';
+import { HashHistory } from './history/hash.js';
+import { HTML5History } from './history/html5.js';
 
 /**
  * @typedef {{
@@ -13,7 +12,7 @@ import { HTML5History } from './history/html5';
 /** @type {Route} */
 let lastRoute = {};
 
-/** @typedef {import('../Docsify').Constructor} Constructor */
+/** @typedef {import('../Docsify.js').Constructor} Constructor */
 
 /**
  * @template {!Constructor} T
@@ -21,12 +20,7 @@ let lastRoute = {};
  */
 export function Router(Base) {
   return class Router extends Base {
-    /** @param {any[]} args */
-    constructor(...args) {
-      super(...args);
-
-      this.route = {};
-    }
+    route = {};
 
     updateRender() {
       this.router.normalize();
@@ -39,7 +33,7 @@ export function Router(Base) {
       const mode = config.routerMode || 'hash';
       let router;
 
-      if (mode === 'history' && supportsPushState) {
+      if (mode === 'history') {
         router = new HTML5History(config);
       } else {
         router = new HashHistory(config);
