@@ -1,12 +1,30 @@
-import { makeExactMatcher } from './exact-match';
-import { createNextFunction } from './next';
+import { makeExactMatcher } from './exact-match.js';
+import { createNextFunction } from './next.js';
 
-/** @typedef {import('../Docsify').Constructor} Constructor */
+/** @typedef {import('../Docsify.js').Constructor} Constructor */
 
 /** @typedef {Record<string, string | VirtualRouteHandler>} VirtualRoutesMap */
 /** @typedef {(route: string, match: RegExpMatchArray | null) => string | void | Promise<string | void> } VirtualRouteHandler */
 
 /**
+ * Allows users/plugins to introduce dynamically created content into their docsify
+ * websites. https://github.com/docsifyjs/docsify/issues/1737
+ *
+ * For instance:
+ *
+ * ```js
+ * window.$docsify = {
+ *   routes: {
+ *     '/items/(.+)': function (route, matched) {
+ *       return `
+ *         # Item Page: ${matched[1]}
+ *         This is an item
+ *       `;
+ *     }
+ *   }
+ * }
+ * ```
+ *
  * @template {!Constructor} T
  * @param {T} Base - The class to extend
  */
@@ -83,7 +101,7 @@ export function VirtualRoutes(Base) {
       }
 
       return {
-        then: function (cb) {
+        then(cb) {
           done = cb;
           asyncMatchNextRoute();
         },
