@@ -2,13 +2,11 @@ function handleExternalScript() {
   const container = Docsify.dom.getNode('#main');
   const scripts = Docsify.dom.findAll(container, 'script');
 
-  for (let i = scripts.length; i--; ) {
-    const script = scripts[i];
-
-    if (script && script.src) {
+  for (const script of scripts) {
+    if (script.src) {
       const newScript = document.createElement('script');
 
-      Array.prototype.slice.call(script.attributes).forEach(attribute => {
+      Array.from(script.attributes).forEach(attribute => {
         newScript[attribute.name] = attribute.value;
       });
 
@@ -22,4 +20,5 @@ const install = function (hook) {
   hook.doneEach(handleExternalScript);
 };
 
-window.$docsify.plugins = [].concat(install, window.$docsify.plugins);
+window.$docsify = window.$docsify || {};
+$docsify.plugins = [install, ...($docsify.plugins || [])];
