@@ -1,4 +1,4 @@
-import parser from './parser';
+import parser from './parser.js';
 
 const install = function (hook, vm) {
   // Used to remove front matter from embedded pages if installed.
@@ -18,4 +18,5 @@ const install = function (hook, vm) {
   });
 };
 
-$docsify.plugins = [].concat(install, $docsify.plugins);
+window.$docsify = window.$docsify || {};
+$docsify.plugins = [install, ...($docsify.plugins || [])];
